Path: First implementation to parse Windows paths
First implementation for parsing Windows paths. The features supported are:
* Normalization of the path (without accessing the file system)
* Getting relative paths
* Appending paths
* Trimming a path (the trailing directory character)
* Getting the parent path

The most useful feature of this library is the high level support of Windows style paths (UNC, DOS, relative) and normalizing the paths to the shortest possible path (interpreting the '..' and '.'), followed by having an absolute path (pinned) and determining the relative path from a base path.

Issue: DOTNET-464