UnixPath: Implementation based on WindowsPath
There is much code in common between the two implementations, but it is decided to keep them separate so that code runs as fast as possible. There are just enough differences between Unix paths and DOS paths that the code would look unnecessarily complicated with a lot more conditional checks. Rely instead on as many unit tests as is possible.

While implementing, identified some minor performance improvements for Append (if the left is empty, take the right).

Issue: DOTNET-467