WindowsPath: Don't allow the use of ':' in the path
The character ":" is not a valid Windows path character. Ensure it is not allowed to avoid potential confusion when parsing paths.

If a path such as `\\server\share\c:\foo` is allowed, and then the relative path to `\\server\share` is requested, this would result in a relative path `c:\foo` which is clearly wrong. This can only be avoided efficiently if the `:` character is disallowed.

Two techniques was tried, using a string search, and checking the character in a loop. The check in the loop is slightly faster.

Issue: DOTNET-473