FileSystemNodeInfo: Allow comparison of two paths if they resolve to the same path
`FileSystemNodeInfo` gets details about a file that can be compared with another `FileSystemNodeInfo` object if the two paths resolve to the same object on the file system. This is regardless of symbolic links.

If paths are not resolved, then the object is of the symbolic link and not the target path.

Test: DOTNETX-8
Issue: DOTNET-852