FileSystemNodeInfo: Fix resolution of symlinks on Linux
`MonoUnixNodeInfo` now raises an exception if it can handle the path, but the path is invalid (as it should have).

`MonoUnixNodeInfo` now uses glibc 6 `realpath` API to resolve the name of the file, instead of trying to do this itself (via `readlink`), which resulted in incorrect behaviour (it didn't take into account the path). Even if we fixed this, any transient symlinks (e.g. as part of the directory path) wouldn't be resolved, which should now be done by the OS. This solution should also be faster (one call, reduced string manipulations). The API `readlink` must be used if `realpath` fails (i.e. the target doesn't exist).

The main class `FileSystemNodeInfo` is simplified in getting the correct OS class.

Test: DOTNETX-20
Issue: DOTNET-852