FileSystemNodeInfo: Fix resolution of symlinks on Windows
The `Win32NodeInfo` now raises an exception for the case that the file information couldn't be obtained, making it consistent with `MonoUnixNodeInfo` and consistent behaviour in case recursing and a file can't be found.

Fixed the case that a junction or a symbolic link was pointing to a non-existent target. To get this information, we have to get the symbolic link details with `DeviceIoControl`. If it's a relative symbolic link, we use our existing `Path` code to get the full path. The code handling non-existent target links are slow, but this should normally be rare (most links would point to valid targets to be useful).

Test: DOTNETX-20
Issue: DOTNET-852