FileSystemNodeInfo: Recurse on Windows XP
Windows XP will raise a `EntryPointNotFoundException` for a junction as the Win32 API `GetFinalPathNameByHandle` is only defined on Windows Vista and later. On Windows XP, we must now recursively use `DeviceIoControl` to resolve the links.

Recursively resolving using `DeviceIoControl` is now also used on Windows Vista and later to find the last unresolved link (instead of just the first), unless recursion is detected, in which the first link is used. Rationale for this behaviour is:

* When recursion occurs, just return the direct link. Thus code can open the link target, it will find that it actually exists and is a reparse point. They can still step by step do the work themselves if they want.
* When recursion doesn't occur, resolve to the very last link that doesn't actually exist.

Secondly, on Windows XP, we have the use case that the junction `PrintName` might not be available, so we also handle the `SubstituteName` properly.

Some minor robustness was added.

Test: DOTNETX-26
Issue: DOTNET-852