FileExecutable (WinExecutable): A mechanism to test if a file is an executable
It may be useful to know if the file is an executable or not, and what Operating System and architecture it is intended for. By using the code snippet:

```csharp
FileExecutable exe = FileExecutable.GetFile(path);
```

we can get the machine type (x86, x64, ARMNT, AARCH64, Itanium64, Alpha), the target OS (e.g. Windows), if it is an EXE, or DLL (it may be neither), Architecture (32-bit or 64-bit), Win Subsystem.

For Windows specifically, get:
* the wordsize (which may be different to the value expected by the machine type, e.g. a .NET file might be i386, but 64-bits).
* the subsystem (Console, GUI, Native, Posix)
* the OS, Image, Subsystem version

.NET Files are not interpreted here. If when, this would be new properties (without faking existing PE methods).

Issue: DOTNET-856