UnixElfExecutable: Read an ELF file
Read a Unix ELF file and determine the architecture, word size, if it's an executable, a dynamic loaded library, position independent, or a core file. It supports various machine architectures (but not all, only those where Mono might have been run). It should be easy to extend however with future machine types.

The tool FileInfo is extended to print information about the ELF file if it is. Note, this tool will not replace `readelf`, it is intended only to read the most basic metadata from an ELF file. MacOS does not use the ELF format.

The implementation is tested against real-world binaries found on the Internet.
* The ALPHA target is observed on earlier versions of Debian to have the machine type 0x9026, although this wasn't found in any (presumably official) documentation.

A specific tool, `FileInfoCheck` is created that allows for integration tests. It is not possible to check only the first 64 bytes of an ELF file. To check if it's an EXE, the p_type must be checked, but also the program headers for P_INTERP. For a DLL, the name should be checked if it contains SONAME. This requires quite a few binaries, which cannot be checked into the repository for a number of reasons:
* The binaries are licensed and copyright
* Some of the licenses don't allow for redistribution (e.g. the media came with the computer, like proprietary Operating Systems, such as Solaris)
* The binaries are too large and should not be checked into GIT (and this repository doesn't use LFS).

Tested against various ELF (exe/dll) binaries for:
* linux-{sparc, hppa, m68k, alpha, i386, amd64, ia64, powerpc, ppc64el, armhf, armel, arm64, s390x, mips, mipsel, mips64el, riscv64}
* openbsd-{luna88k, octeon, riscv64, alpha, sparc64, amd64, hppa, i386, macppc, powerpc64}
* netbsd-{acorn32, alpha, amd64, amiga, dreamcast, earmeb, earmv6hf, earmv7hf, i386, ibmnws, sparc, sparc64, vax, zaurus}
* qnx650-{armle, armle-v7, mipsbe, mipsle, ppcbe, ppcbe-spe, shle, x86}
* qnx710-{aarch64le, armle-v7, x86_64}
* solaris-{sparc64, x86_64}

Issue: DOTNET-857