FileSystemNodeInfo: The Path property should rely on the file system
The classes `Win32NodeInfo` and `MonoUnixNodeInfo` should now get the `Path` property from the Operating System once it has been resolved.

On Windows, this solves a difficult problem if the user gives a drive letter (that is different to the current drive letter). The actual current directory is unknown without changing drive letters (Windows maintains a current directory per mapped drive). Therefore, it is impossible to know the full path.

Further, giving say the argument `X:` on Windows would raise an exception because of a bug if the current working directory was `C:\xx`. Even if that could be fixed easily, the problem with trying to append across drives is not easily solvable. So we let Windows do it.

On Windows Vista and later, the solution is easy - just use `GetFinalPathNameByHandle`. On Windows XP this doesn't exist, so some heuristics is done using `NtQueryInformationFile` and `NtQueryObject` is used to get the file name. The NT paths are converted to DOS paths where possible.

Fixed minor bug in GenericNodeInfo that would have a hash that was different if files were equal when case sensitivity is disabled.

Internally, exceptions aren't raised everywhere any more, use the `Result` from RJCP.Core.Datastructures. This has a minor impact that locations where the errors occur isn't available with the currently supported .NET runtimes (NET 4.x, .NET Standard 2.1). Would need to upgrade to .NET 6.0 or later.

Issue: DOTNET-866