Build: Upgrade to MSBuildTasks v0.2.6
Version v0.2.5 would set the assembly version to a pre-release version, even if the project provided an override. So that documentation builds with the correct version (and not a prereelase), v0.2.6 doesn't use a pre-release tag if the git tag has an override in the master project `.rjbuild` file.

Issue: DOTNET-1072